Improvement: Reorganized Faction Standing Package
